requirements.txt
# Core dependencies
numpy>=1.24.0
sounddevice>=0.4.6
soundfile>=0.12.1

# Wake Word Detection
pvporcupine>=3.0.0

# Voice Activity Detection
webrtcvad>=2.0.10

# Speech-to-Text
faster-whisper>=0.10.0
torch>=2.0.0

# MQTT Client
paho-mqtt>=1.6.1

# Text-to-Speech (Piper)
piper-tts>=1.2.0

# Configuration
PyYAML>=6.0

# Audio processing
pyaudio>=0.2.13
